Set the VlocityDataPackType when there's a compilation error
At the moment, if we try to deploy a SASS file that doesn't compile, the JSON output doesn't include VlocityDataPackType.
The JSON output does include an entry for the data pack that couldn't be deployed.
I'm parsing this JSON output and want to include the VlocityDataPackType for SASS files that don't compile.

We can reproduce this by deploying a VlocityUITemplate with a .scss file that has an import referencing a file that doesn't exist in the deployment.
In the `projectPath` directory in the job file, we create a data pack `VlocityUITemplate\campaign-base-grid` with a `campaign-base-grid.scss` file.
At the start of `campaign-base-grid.scss` file, we have `@import "campaign-variables";`, which doesn't exist in a deployment.

When we run `vlocity packDeploy -job DataPackJob.yaml -propertyfile source.properties -json` to deploy the data pack, the current output for this data pack (in the `records` property of the JSON) is:
```
        {
            "VlocityDataPackKey": "VlocityUITemplate/campaign-base-grid",
            "VlocityRecordSObjectType": "%vlocity_namespace%__VlocityUITemplate__c",
            "Name": "campaign-base-grid",
            "VlocityRecordSourceKey": "%vlocity_namespace%__VlocityUITemplate__c/campaign-base-grid",
            "VlocityDataPackStatus": "Error",
            "VlocityDataPackDisplayLabel": "campaign-base-grid (VlocityUITemplate/campaign-base-grid)",
            "ErrorMessage": "SASS Compilation Error Failed to compile SCSS: {PROJECT_PATH}\\VlocityUITemplate\\campaign-base-grid\\campaign-base-grid.scss SASS compilation failed, see error message for details: Error: \n        on line 2:9 of /stdin\n>> @import \"campaign-variables\";\n\n   --------^\n"
        }
```

With this change, the output has the `VlocityDataPackType` key:
```
        {
            "VlocityDataPackKey": "VlocityUITemplate/campaign-base-grid",
            "VlocityRecordSObjectType": "%vlocity_namespace%__VlocityUITemplate__c",
            "Name": "campaign-base-grid",
            "VlocityRecordSourceKey": "%vlocity_namespace%__VlocityUITemplate__c/campaign-base-grid",
            "VlocityDataPackType": "VlocityUITemplate",
            "VlocityDataPackStatus": "Error",
            "VlocityDataPackDisplayLabel": "campaign-base-grid (VlocityUITemplate/campaign-base-grid)",
            "ErrorMessage": "SASS Compilation Error Failed to compile SCSS: {PROJECT_PATH}\\VlocityUITemplate\\campaign-base-grid\\campaign-base-grid.scss SASS compilation failed, see error message for details: Error: \n        on line 2:9 of /stdin\n>> @import \"campaign-variables\";\n\n   --------^\n"
        }
```