Fixes an issue with exporting Products, etc with hyphens.
This fixes an issue with exporting a String Translation (and possibly other items), where the Name is a single hyphen. The data sanitization removes the hyphen, leaving the resulting folder name as an empty string, and breaking the export. This ensures that if the sanitization removes the full string, that there is a replacement.