Disable colouring of SFDX CLI output (#560)
* Update readme instructions --skip-ci

* Code Delivery 1.16.1

* Disable Extensive Logging & Repetitions on Git Based Deploy (#523)
* Fixes local compilation #525 (#532)

* Improve the documentation for lwcCompiler (#528)

lwcCompilerVersion and lwcCompilerNamespace documentation should be clearer. Currently the example is using the release version such as 236 or 238, but it needs to be the package version such as 900.469. Also examples of namespaces would be good.

* Disable colouring of SFDX CLI output

Disable colouring of SFDX CLI output so it can be JSON parsed.
Else the output of source:convert will look like this, and cannot be processed

---------

Co-authored-by: Adam Rutland <arutland@salesforce.com>
Co-authored-by: Manas Agarwal <manas.agarwal@salesforce.com>
Co-authored-by: Dean Hobden <dean.hobden@centrica.com>